brew-build-bottle-pr: browse PR only when necessary ENV vars are set
`hub pull-request` can browse created PRs when you pass `--browse`. This command, however, uses BROWSER or HOMEBREW_BROWSER env. vars to determine which program to use for browsing. When none of them are set, `hub pull-request` fails. Here we modify the code so that `--browse` is invoked only if at least one of BROWSER or HOMEBREW_BROWSER is set.

Signed-off-by: Maxim Belkin <maxim.belkin@gmail.com>